fixed problem with running unit tests on 64 bit machine
git-svn-id: https://svn.urbansim.org/trunk@11456 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/samplers/stratified_sampler.py
@@ -278,7 +278,7 @@ class stratified_sampler(Sampler):
 
 
 from opus_core.tests import opus_unittest
-from numpy import all, alltrue, not_equal, equal
+from numpy import all, alltrue, not_equal, equal, int32
 from opus_core.datasets.dataset import Dataset
 from opus_core.samplers.constants import *
 from opus_core.storage_factory import StorageFactory
@@ -335,7 +335,9 @@ class Test(opus_unittest.OpusTestCase):
                                         self.households.get_attribute("grid_id")[index1],UNPLACED_ID)
                 chosen_choice_index = UNPLACED_ID * ones(index1.shape, dtype="int32")
                 w = where(sample_results[1]>=0)[0]
-                chosen_choice_index[w] = sampler[w, sample_results[1][w]]
+                # for 64 bit machines, need to coerce the type to int32 -- on a
+                # 32 bit machine the astype(int32) doesn't do anything
+                chosen_choice_index[w] = sampler[w, sample_results[1][w]].astype(int32)
                 assert alltrue(equal(placed_agents_index, chosen_choice_index))
                 sampler = sampler[:,1:]
             assert alltrue([x in index2 for x in sampler.ravel()])
@@ -367,7 +369,7 @@ class Test(opus_unittest.OpusTestCase):
                                         self.households.get_attribute("grid_id")[index1],UNPLACED_ID)
                 chosen_choice_index = UNPLACED_ID * ones(index1.shape, dtype="int32")
                 w = where(sample_results[1]>=0)[0]
-                chosen_choice_index[w] = sampler[w, sample_results[1][w]]
+                chosen_choice_index[w] = sampler[w, sample_results[1][w]].astype(int32)
                 assert alltrue(equal(placed_agents_index, chosen_choice_index))
                 sampler = sampler[:,1:]
             assert alltrue([x in index2 for x in sampler.ravel()])
